requirements.txt
fastapi
uvicorn
openai
azure-search-documents
python-dotenv
pydantic
